+++ libraries/SRF05/SRF05.h
@@ -2,7 +2,7 @@
 //
 //    FILE: SRF05.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.3
+// VERSION: 0.1.4
 //    DATE: 2021-05-17
 // PURPOSE: Arduino library for SRF05 distance sensor
 //     URL: https://github.com/RobTillaart/SRF05
@@ -10,7 +10,7 @@
 
 #include "Arduino.h"
 
-#define SRF05_LIB_VERSION                 (F("0.1.3"))
+#define SRF05_LIB_VERSION                 (F("0.1.4"))
 
 
 const uint8_t SRF05_MODE_SINGLE      = 0;
@@ -30,12 +30,12 @@ public:
   float    getSpeedOfSound();
 
   //  adjust timing
-  void     setCorrectionFactor(float factor = 1);
+  bool     setCorrectionFactor(float factor = 1);
   float    getCorrectionFactor();
 
 
   //  operational mode
-  void     setModeSingle();
+  void     setModeSingle();  //  default
   void     setModeAverage(uint8_t count);
   void     setModeMedian(uint8_t count);
   void     setModeRunningAverage(float alpha);
@@ -52,13 +52,18 @@ public:
 
 
   //  Experimental - calibration
-  float    determineSpeedOfSound(uint16_t distance);
+  float    determineSpeedOfSound(uint16_t count);
 
 
-  //  Experimental - adjust trigger length - gain is a few usec at best.
-  //  10 microseconds is advised minimum
-  //  void     setTriggerLength(uint8_t length = 10) { _triggerLength = length; };
-  //  uint8_t  getTriggerLength() { return _triggerLength; };
+  //  Experimental - adjust trigger length 
+  //  - gain is a few microseconds at best.
+  //  - 10 microseconds is advised minimum
+  //  - to be investigated.
+  void     setTriggerLength(uint8_t length = 10);
+  uint8_t  getTriggerLength();
+
+  //  TIMING
+  uint32_t lastTime();
 
 
 private:
@@ -72,11 +77,12 @@ private:
   float    _correctionFactor = 1;
   uint8_t  _triggerLength    = 10;
   float    _speedOfSound     = 340;      //  20°C
-
+  uint32_t _lastTime = 0;
+ 
   uint32_t _read();
   void     _insertSort(uint32_t * array, uint8_t size);
 };
 
 
-// -- END OF FILE --
+//  -- END OF FILE --