+++ libraries/SRF05/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.4] - 2023-03-15
+- add **uint32_t lastTime()**
+- changed return type **bool setCorrectionFactor()**
+- add experimental **get- setTriggerLength()**
+- redo **determineSpeedOfSound()**
+- update readme.md
+- update GitHub actions
+- update license 2023
+- minor edits
+
+
 ## [0.1.3] - 2022-11-25
 - Add RP2040 support to build-CI.
 - Add CHANGELOG.md
@@ -14,7 +25,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add mode constants
 - update unit test
 - move code from .h to .cpp
-
 
 ## [0.1.2] - 2021-12-28
 - update library.json